Initial values for alien strategy.
Our strategy has chances for each region to be selected, and for each mission
type in each region.
We are loading initial values from the ruleset ('startingStrategy') and keeping
it in the save file('alienStrategy'). But we are not actually using it yet.

The values are taken from the ufopaedia.org pages for ZONAL.DAT and ACTS.DAT.
We don't have values for retaliation, terror and supply missions, they (will)
get special treatment in the code.